Updated openstack/openstack Project: openstack/cinder  7ab30f99a5877919d0fb81274a21b605064b3f6c
Update Pure replication to cheesecake

Migration to cheesecake replication. This basically
modifies any of the existing replication code to work on the whole
backend instead of the specified volume.

It also completes the transition to allow swapping the underlying array
being managed by cinder upon failover. In the v2 implementation it was
conditional on some operations where things *could* be on the other
array. Now it is all driver operations.

This also switches the target_device_id to backend_id in the config for
replication devices.

Change-Id: I7f2580f4d764266b62a755cb4529c290e639984c
Closes-Bug: #1542100