Updated openstack/openstack Project: openstack-infra/subunit2sql  b7b7979cf363e4b30aa2fe6576634df48aaec75a
Don't extend the passed in targets

If the caller passes in a value for targets, copy it before extending
it with our own targets.  Because the argument is a reference, our
extending the list here may cause the list to be extended again if
the caller creates another ReadSubunit instance reusing the same
argument.

This caused "ValueError: I/O operation on closed file" errors because
the pares_outcome callback of every previous instantiation was being
called for each new outcome.

Change-Id: Iadd7b063e61590869dbea6ce9db9ffe611fbae0a