Updated openstack/openstack Project: openstack/neutron  2d51a534c2c6f38234d88255fb12b9183a9ff1fc
make/update_subnetpool_dict call _dict_extend

Now the extension fields were not being added to the response for
subnetpool, and when update subnetpool, the response of updated was not
contain any extensions which based on 'standard_attr'.

This patch add 'call _extend_dict' functions to make subnetpool support
for extending dict.

Change-Id: I7bc449aa079d84869b2a707b36bbbfb22d540cc5
Closes-Bug: #1548161