Updated openstack/openstack Project: openstack/python-cinderclient  a330ff0bba225a63379010786c1dce131554fd37
Add --cascade to volume delete

This adds the --cascade parameter to the volume
delete command.  Using this will call the volume delete
with an additional parameter that deletes snapshots as
part of the same operation.

The name "cascade" is chosen to be somewhat generic,
so that we can extend this if desired in the future.
(And to hint at the terminology from SQL.)

Depends-On: I33d15b76d4bd0de14c635d404b2c97096c977a58
Change-Id: I05ce61e647c43f9a1e6d74444356a76e5d284673
Blueprint: del-vols-with-snaps