Updated openstack/openstack Project: openstack-infra/shade  63139318490521220fe41def96d516f347d01a3e
Fix create_server() with a named network

If 'network' was supplied to create_server() along with an empty
'nics' list, we would never attempt to find the network. The os_server
Ansible module uses the API this way. This treats an empty list the
same as if the 'nics' parameter were never supplied.

Change-Id: Idc844fab2c4c08f158c892104d065e5554ed90f3