Updated openstack/openstack Project: openstack/os-cloud-config  ffd214fdcccf58328efafa69d101f46967e4a3bb
Remove admin role in admin tenant from ceilometer

We don't need to grant the ceilometer user the admin role in the admin
tenant any longer. The admin role in the service tenant is sufficient,
as long as ceilometer is configured to use the service tenant. The
Depends-On patches make the switch to the ceilometer configuration to
use the service tenant.

As part of this change, the ResellerAdmin role is also added to the
ceilometer user in the service tenant. Previously, the ResellerAdmin
role was created, but never granted to the ceilometer user.

This brings the ceilometer configuration in line with the
recommendations from the official installation guide:
http://docs.openstack.org/kilo/install-guide/install/yum/content/ceilometer-controller-install.html

Change-Id: I19f7b3aa156ff879fd1c9521bc9e172c27a3d6b7
Partial-Bug: #1358237
Depends-On: I40dd4543580332a46cde96f4bca639eb777ac7a8
Depends-On: Ia14695eb23a1ff551fd27f74b4cb864e80b100e3