Updated openstack/openstack Project: openstack-infra/shade  f39105176d47e517458d98ca9a31e89fd2abc943
Add test for os_server Ansible module

We haven't been able to test os_server because we haven't had a
way to reliably determine the image name (which changes from time
to time based on version) from the playbook. A stupidly clever
hack is to just grep out the name from a shell call to the openstack
client and pass that in to the playbook from the command line.

The playbook is simple for now for the initial commit. I expect new
tasks to be added to it later.

Change-Id: I3caedfd4c805126c8da1fc8bab06ccda3fd96528