Updated openstack/openstack Project: openstack-infra/project-config  d45c220b14062ea6d4ea4aa0a9d7a7429ee5e149
Add a third scenario for packstack integration testing

We have taken measures to both further split scenarios [1] and
introduce RAM saving features [2] not unlike the work that was
recently done for puppet-openstack scenarios so that jobs can
work within the memory constraints.

We would like to add scenario003 to the check non-voting
pipeline to complete our verifications. Once all jobs are stable,
we will set them all to voting and add them to the gate.

[1]: https://review.openstack.org/#/c/285516/
[2]: https://review.openstack.org/#/c/286158/

Change-Id: Ieac60ee40a9620d330b17f88ba89146d6cc4c689