Updated openstack/openstack Project: openstack/swift  48b5e8a13fe7ae6a83e8beef773b1d28b2775b29
Fix bug in mem_diskfile write_metadata method

The mem_diskfile DiskFile.write_metadata method was
apparently never called in any existing test, as if it
were it would blow up as reported in the bug.

This patch fixes the method and adds a test that
exercises it. The test addition itself should be useful
since it verifies the behaviour of Last-Modified after
POSTs to an object.

Drive-by fixes for bad docstring and undefined references
in the _quarantine method.

Change-Id: I17fd62e5f02be5b48bfd9ba7fa25315e30a0a4bf
Closes-Bug: #1536037