Updated openstack/openstack Project: openstack/nova  32051990e9b2855ea48e03b36e70d8c83a6d7a5f
objects: Allow instance to reset the NUMA topology

This will be used when aborting claims, since the claim will (if
successful) update the instance with the claimed topology,

In case of an abort - we want to make sure it's clear for consistency
sake. This will be done in the follow up patch.

Change-Id: I2989446fdaa44a30d90ae2ab29fc27fb2ad03c4a
Partial-bug: 1545675