Updated openstack/openstack Project: openstack/horizon  dcc838128e0812a760733f24a65f09bcd047a399
Fixing heatclient release compat issues

The new release of python-heatclient broke tests that were using
hardcoded sample templates. For the json based templates, the new
line was escaped. For yaml templates, a comment section was added to the
templates.

My working theory is that the exception handling changed and we were
missing an error before that is being exposed now. That is to say, our
hardcoded templates were wrong before, we just weren't seeing it.

Tested with latest version of python-heatclient 1.0.0 and last version
0.9.0.

Closes-Bug: #1552400
Change-Id: Ibdf985654ebfa60205068b167a37600f7ed4c1f4