Updated openstack/openstack Project: openstack/oslo.versionedobjects  788e3d08f7d073acb9a9958b4e474c9d51da934d
Use primitive type name in ValueError for Object field coerce failure

When the Object field coerce method doesn't get an object, the
type put into the ValueError message is the empty string, which is
not useful.

This change uses the type name from the value so the error has
useful details.

Change-Id: I45386cfda17243a82864780da1701d3b35432ae0
Closes-Bug: #1552480