Updated openstack/openstack Project: openstack/cinder  97b2df7ad7150eb29754670c1204f32521a6253f
Huawei: Manage volume fails due to lower array version

When we try to import an existing volume to cinder, we try
to find whether the volume already exists in a hypermetro.
But if the array version is lower than V3R3, the request
to find the hypermetro info will fail because array version
lower than V3R3 does not support hypermetro.

Closes-Bug: #1546375
Change-Id: Ia28e62f3924155e813df80f3c4932e3f0f10333e