Updated openstack/openstack Project: openstack/nova  23fdf28daf49d8bf0033d023fb0c73cd69ae5314
Deprecate nova.hooks

This emits a warning message when hooks are used to ensure that admins
understand that this facility is deprecated in Nova going forward. It
also removes the links to the meager documentation on hooks to ensure
no one finds these by accident and starts using them. Also add a
comment to the code file itself incase someone ends up in it from
other searches.

Change-Id: I31b1223a0dc3fdffe356a816b591fdd88a31483a