Updated openstack/openstack Project: openstack/python-neutronclient  2e0dc1f469af37cab2eea6be7b4c801c70c04eaf
Support dry-run option for auto-allocated-topology

Add the ability to pass fields with the
auto-allocated-topology-show command to support the
dry-run validation option of the API.

With dry-run the CLI result is "Pass" or the error message
from the response.

Rename the client binding from show_ to get_. Also provide a
client binding for validating the auto-allocation requirements.

Partially-Implements: blueprint get-me-a-network

DocImpact: Add info about dry-run to the auto-allocate section
           in the networking guide.

Change-Id: Ieba6f3cde23a8a93067b8239b096d5103f6a3128