Updated openstack/openstack Project: openstack/barbican  598d3b89a09e078730b246b6c7567193a4c8a235
Adding check on Offset

When a large offset is specified in GET API, causes
failure with 500 error. Adding check if offset is
greater than sys.maxsize, set offset to 0 instead of
failing with 500. Setting offset to 0, returns the
number of records (as specified by limit) ordered by
created_date.

Comparing offset with the sys.maxsize which is
compatible of all the different python versions available.

Closes-bug: #1481012

Change-Id: I70a59e0edb9378271e62b6759dfb02c8c0110ac0