Updated openstack/openstack Project: openstack/neutron  230068bf8544355d999bae2bfd6e35cbfebbb2a7
DVR:Remove unwanted check in _get_dvr_service_port_hostid

There is a check in _get_dvr_service_port_hostid to verify
if the port belongs to dvr service port or an agent gateway
port.

There is no need to check for the agent gateway port host.
This might be a left over from the previous cycle cleanup.

This patch removes the unwanted check.

Change-Id: I2571ec3453c2a5cc95c17b9c65de482e7e0ce9e5