Updated openstack/openstack Project: openstack/tempest  8eb795b08fe38d661d2008e23deb4a5eb9135731
Make object/orchestration clients use rest_client

Now service_client is unnecessary because that is the same as
rest_client. This patch makes object_storage and orchestration
clients use the rest_client, and we will remove service_client
for the cleanup in the future.

This patch removes these clients from a certain test because
this test is just for service_client, not rest_client as the
test name. We have added this test for verifying all arguments
are passed into rest_client via service_client. However, this
patch some clients use rest_client directly, so this test becomes
meaningless for these clients. That is a reason why the migrated
clients(into tempest.lib before) are not contained in this test.

Partially implements blueprint consistent-service-method-names

Change-Id: I763107de6a9f7b411a5892d2c53dcf99b7afa3ef