Updated openstack/openstack Project: openstack/nova  a3cf38a3ec0fd57679320688bd815225c2bf053f
Abstract a driver API for triggering crash dump

An implementation to trigger crash dump depends on hypervisors. Inject_nmi is
just a libvirt implementation. The driver API name should be
"trigger_crash_dump" for the abstraction.

This patch only changes names of the driver API and an Exception. It has no
functional changes.

Change-Id: Icef57d346fac7f92132bc14e22d5b2652397a64a
Closes-Bug: 1547362