Updated openstack/openstack Project: openstack/tripleo-heat-templates  f8e58b621244a5873115510fca7ec2f66aac4958
Add Satellite 5 support

Add Satellite 5 support to the RHEL registration environment and
resources. The registration script is updated to support both satellite
versions in place given the similarity of the options for both
scenarios.

The satellite version is detected based on $REG_SAT_URL, and that
determines whether subscription-manager or rhnreg_ks is used.

Change-Id: Ic261c8a16a7d6d3978f8bfc6e53f75dbe1b716db