Updated openstack/openstack Project: openstack/diskimage-builder  6e29b45d90ed37f47bbd121be135207e6a0a7fa5
Replace sfdisk partitioning with parted

A TODO was placed on the partitioning section of the vm element to
replace sfdisk with a saner (and less arcane) way of partitioning. It
suggested parted for replacement. This changeset should reproduce the
same disk label and partition layout as sfdisk, but with less ioctl
errors and version dependency. It will also ensure partition alignment.

Change-Id: I5d8d75131458b73bfb05f80f1bfa7e2970e004b3