Updated openstack/openstack Project: openstack/nova  fe03930eea8e4fabece5b1c1732fef783ce4c402
Ensure there are no unreferenced closures in tests

In order to add a hacking check which ensures that closures defined in
tests are actually used a few instances where they are not must be
cleaned up. This removes a few that were not used and uses a few that
should be used.

Change-Id: Ifb76525b96de49eccda5311c5d96c7644d99336b