Updated openstack/openstack Project: openstack/cinder  11be8173564933d4f16182d7a7e750362f27a0c5
Fix oslo.service config generation

Cinder generation configuration sample does not contain any of
oslo.service configuration options.

This is because we are not using the right namespace when calling
oslo-config-generator since oslo.service is not a valid entrypoint as
mentioned in the documentation:

https://github.com/openstack/oslo.service/blob/master/doc/source/usage.rst#using-osloservice-with-oslo-config-generator

This patch fixes this by using the right namespaces.

Closes-Bug: #1552689
Change-Id: I8365daec0d3b161d3c07f229d221cf004edb6662