Updated openstack/openstack Project: openstack-infra/system-config  1b7f729f73c7cfdb618d9407d73dd1ae66bdce58
Add a mailing list for the French i18n local group.

The French i18n team is getting more and more members and
Zanata lacks an efficient mailing feature. So the solution
is to use a mailing list.

In agreement with the i18n PTL, the ML is called openstack-i18n-fr.

Change-Id: I35079e25cd03b02dbbf5444acf5d574e51508dfd