Updated openstack/openstack Project: openstack/tempest  e257898517262b01dbd44e52cd378a2929bf5b88
Tokens need user domain be created correctly

When calling get_token() there are many assumptions on the
domain when user_domain_id or user_domain_name and
project_domain_id or project_domain_name are not specified.
This is okay when running tempest against devstack, since
the user is in the default domain but in productions the
test user is rarely in the default domain. Instead we should
also pass user_domain_id or user_domain_name and
project_domain_id or project_domain_name to
get_token to create a token.

If there is a case where user_domain_id is None when passed
to get_token() the previous functionality still takes into effect.
Assuming the user is in the 'Default' domain.

Change-Id: I78c472f2284a8641c0c7cf0b3e0994984a04c5b0
Closes-Bug: #1548987