Updated openstack/openstack Project: openstack/nova  410377ab4bc1f423bd33c6589517deb2ba303468
XenAPI: introduce unit test for XenAPI plugins

This is the first patch set to introduce unit test for XenAPI
plugins. It contains the common part depended by other plugin
tests. In order to make this first patch set to be as simple
as possible, it only covers the test framework and the unit
test for a simple plugin.

Change-Id: Iece439661cc4337e826d0eb15b438d2716214fb5