Updated openstack/openstack Project: openstack/ironic  036b959b9245ba44b5ea895ea48584dedc282fdf
Fix unittests after new releases of libraries

ironic-lib and galnceclient releases caused some of unittests to fail,
this change fixes them.

Two unittests in the test_glance_service.py module was depending on
getting an instance of the real glanceclient and it was failing due a
new glanceclient release. This patch is fixing the problem by mocking
the client instanciation.

In case of ironic-lib, new argument was added to make_partitions
function, and the mocks need to be changed.

Closes-bug: #1552839
Closes-Bug: #1552834
Co-Authored-By: Lucas Alvares Gomes <lucasagomes@gmail.com>
Change-Id: Ic9dc3f1016667c6d7ec5ee324d5fff48af82c581