Updated openstack/openstack Project: openstack/nova  e33f8b39c0922584e785732b94c6a5dbdcaffeba
Create BuildRequest object during boot process

The BuildRequest object represents a boot request before an instance has
been created and written to the database. It will be required to respond
to API requests to list/show an instance in cellsv2 where the instance
is not written to the database immediately. An upcoming change will
modify the boot process to defer instance creation until after
scheduling has picked a cell/host and having this BuildRequest object
persisted is a prerequisite for that work. The BuildRequest object
lifetime ends as soon as the instance has been written to a database and
a mapping is in place for it. Until instance creation moves the
BuildRequest object is destroyed rather quickly.

Note there is a small optimization added here such that when creating
instances in the API, only populate the SecurityGroupList once rather
than each pass in the for loop (if building multiple instances).

Change-Id: I080fa53d13f7301f6f7f2977a8e3cb287a46c02c
Partially-implements: bp cells-scheduling-interaction