Updated openstack/openstack Project: openstack/nova  2c732aa654c3ce1be7f04b557720de432b1b1578
Make InstanceMappings.cell_id nullable

In order to indicate that an instance boot request has been received,
but the instance has not yet been scheduled to a cell, it should be
mapped to a null cell_id. This will mean that an instance list/show will
need to fetch the instance details from the BuildRequest object rather
than looking for the instance in a cell database. A migration is added
and the InstanceMapping object is modified to make this possible.

Partially-implements: bp cells-scheduling-interaction

Change-Id: Iee768512314f11a13e2a85a1e7d5442864ddbc4b