Updated openstack/openstack Project: openstack/ironic-python-agent  9b9b982fddd690243d435ebdeadc4186edd951d9
Wait for udev to settle before listing the block devices

This patch is making the list_all_block_devices() method to wait for
udev to settle it's event queue prior to listing the devices.

Sometimes the ironic-python-agent service may start before all devices
were detected and end up erroring out because it couldn't find a
suitable disk for deployment.

Closes-Bug: #1551300
Change-Id: I1ae2062a711115a1ea14b79ae9ace7ddd2fff9d5