Updated openstack/openstack Project: openstack/neutron  9c5c121a18059a2c80a6e0be7607d7d0584c2543
Fix pecan collection->plugin map for extensions

The way pecan was mapping collections to plugins always assumed
that the same plugin that supported an extension that extended
a given collection was responsible for the collection itself.
This isn't the case for extensions that just add data to a
collection via the dictionary extension mechanisms. For example,
the timestamps are implemented in an extension that is run by
a standalone service plugin that just performs DB updates via
DB hooks and gives the responses to users via the dictionary
extension hooks.

This patch fixes it by checking to ensure the plugin actually
supports the get method for the collection itself before assuming
that the collection belongs to the plugin.

Closes-Bug: #1552926
Change-Id: I22fa396e1ccd6af936fd4d58bbb0638ff5ab4467