Updated openstack/openstack Project: openstack/cinder  0cc683951079e9cf726f02bee47400e166abdbbc
3PAR use same LUN id for each export path

When multipath is enabled, and the 3PAR driver is configured
to export a VLUN over N iSCSI IP(ports), we now use the same
LUN ID for each iSCSI IP export.  Previously, when we would
export a volume over N iSCSI ports, we would get a new LUN ID
for each iSCSI port.

Change-Id: Ide0ce373a14f348a554688dee9b699feed3c5f26
Closes-Bug: #1551994