Updated openstack/openstack Project: openstack/tempest  c3ccd11d170d19add106fec6b71cd7bbe7e4b184
Make volume clients use rest_client

Now service_client is unnecessary because that is the same as
rest_client. This patch makes volume clients use the rest_client,
and we will remove service_client for the cleanup in the future.

This patch removes these clients from a certain test because this
test is just for service_client, not rest_client as the test name.
We have added this test for verifying all arguments are passed into
rest_client via service_client. However, this patch some clients use
rest_client directly, so this test becomes meaningless for these
clients. That is a reason why the migrated clients(into tempest.lib
before) are not contained in this test.

Partially implements blueprint consistent-service-method-names

Change-Id: I1f7d8bfdb23500342984508333321422f8e2366b