Updated openstack/openstack Project: openstack/tempest  523033a396ceb1026e90c2128cb2035db17c3e1b
Make image clients use rest_client

Now service_client is unnecessary because that is the same as
rest_client. This patch makes image clients use the rest_client,
and we will remove service_client for the cleanup in the future.

This patch removes these clients from a certain test because this
test is just for service_client, not rest_client as the test name.
We have added this test for verifying all arguments are passed into
rest_client via service_client. However, this patch some clients use
rest_client directly, so this test becomes meaningless for these
clients. That is a reason why the migrated clients(into tempest.lib
before) are not contained in this test.

NOTE: This patch changes __init__() of images_client to use
      **kwargs because service_client removed them if each value
      is None, but rest_client doesn't do it. The purpose of the
      __init__() is just for calling self.skip_path(), so this
      change is more readable also.

Partially implements blueprint consistent-service-method-names

Change-Id: I6655b31239d4f3d0c4601b875a4a26778fc9dd2d