Updated openstack/openstack Project: openstack/nova  7083166966300d9e138eb5f87d3f76ff8c532367
neutron: handle attach interface case with no networks

It's possible to boot an instance without a network and it's
also possible to try and attach an interface without specifying
a port, network or fixed IP. If no network is specified on the
attach request and there are no networks available to the
instance's project (shared or not), then the
allocate_port_for_instance method fails with an IndexError.

The code is currently checking for a case of ambiguous networks
when a specific network ID is not requested, but is not checking
for the case that no network ID is specified and no networks are
available. This change adds that check.

A new type of exception is raised from the network API and handled
in the REST API so we return a 400 to the user rather than a 500.

Note that we don't return an empty network_info to the compute
manager because that results in InterfaceAttachFailed raised to
the REST API which is interpreted as a 500.

Closes-Bug: #1545117

Change-Id: Iad762ebef08c259339ea5582e65266620fbab0ac