Updated openstack/openstack Project: openstack-infra/storyboard-webclient  a351e1864bf1c147bd4345f82ad36d7156f29030
Allow a custom format string to be passed to time-moment

This adds the ability to pass a custom format string to the
directive to allow the format to easily be different in different
places.

It also uses the formatting filter provided by angular-moment, so
format strings passed to the directive need to be in the format
expected by momentjs [0].

Finally, this patch stops showing "time-ago" format dates for dates
that are in the future.

[0]: http://momentjs.com/docs/#/displaying/format/

Change-Id: I919e8f1fd94b86e6305f44918101c9c9a62482e9