Updated openstack/openstack Project: openstack/requirements  ee9dc0870fa9a9c0c659cfca4dc9e99a08a52359
Add couchdb to global requirements

Update global requirements to get couchdb library. This
is required for running the integration tests in:
https://review.openstack.org/#/c/270443/

The python library for couchdb is maintained at:
https://pypi.python.org/pypi/CouchDB/1.0
This library is python 3 compatible. It carries the BSD license as well
as OSI approved. The code can be found at:
https://github.com/djc/couchdb-python/ and the testing results can be
viewed at: https://travis-ci.org/djc/couchdb-python. This is a new
requirement and the global-requirements.txt does not contain any
other libraries that covers any part of this functionality.

Change-Id: Iecb4552c50dfec3a5b31f60b0f1739eb4eb6e67a