Updated openstack/openstack Project: openstack/nova  6b20476b63cd54a87932d00696008f5227864b25
deprecate manager class options

We should deprecate the manager classes used for all of our
services. This is not a thing we actually expect or want to support
people replacing. If we want modular plug points at any of these all
the options should be in tree and specified by constants to switch
between.

Change-Id: Id25bd4870c6e2fda08dc0177b7ed61a8a6091838