Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  50c21b8c3f7d98d73bd7b3fe1fed1232867b816a
Change /etc/logstash dir ownership to root

This commit updates the /etc/logstash config dir (where the
subunit2sql configs live, which makes so much sense) to be owned by
root. There is no logstash user ever created on the worker nodes so
we shouldn't try to assign it as the owner.

Change-Id: Ifd85cf4d33c85a958aa33fca50bdd847a651e560