Updated openstack/openstack Project: openstack/cinder  dc4761b36bf4fb72d9461781948a254f5acc836e
Convert huawei ISCSIDriver unit tests to RFC5737 addrs

In the course an earlier review [1] it was suggested that
it would be good to change the live IPv4 addresses in these
tests to RFC5737 addresses [2], a range of addresses dedicated
for documentation, unit tests, and the like.

This commit makes these address substitutions without impacting
the function of the unit tests themselves.

[1] https://review.openstack.org/#/c/282861/
[2] https://tools.ietf.org/html/rfc5737

Change-Id: Ice065076010759f69dcb9a1ec06324f4588f202f