Updated openstack/openstack Project: openstack/requirements  6ac77f7d96b21b86216d84cc3bb7ca21e8cb71bb
Add Freezer modules and projects for Mitaka release

Added the following modules to global-requirements.txt
and upper-contraints.txt:

- apscheduler
Is the library actively maintained?
Yes, last version releaset 2015-12-10 as per
    - https://pypi.python.org/pypi/APScheduler/3.0.5

Is the library good code?
The library provides:
    Tests:
    - https://bitbucket.org/agronholm/apscheduler/src/a6545cf29831?at=master
    Extensive documentation:
    - http://apscheduler.readthedocs.org/en/3.0/
    Active bug tracking:
    - https://bitbucket.org/agronholm/apscheduler/issues?status-new&status-open

Is the library python 3 compatible?
    - It support ptyhon>=3.3

Is the library license compatible?
    - Library license is MIT, thus a compatible license

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?
    - Debian:
        - https://packages.debian.org/jessie/all/python-apscheduler/download
    - Fedora:
        - https://admin.fedoraproject.org/pkgdb/package/rpms/python-APScheduler/

Is the function of this library already covered by other libraries in global-requirements.txt
    - As far as we can see, there's not other scheduling related module available.

Change-Id: Iba93963f5e633f3711b26e7eed16a5b623e7683d
Depends-On: Iae75d00f6b3d18c55b32fd11b91836a1ce4ddd95