Updated openstack/openstack Project: openstack/cinder  ad1d1034e06315ff1812f592c2ce42781cd31c15
Fix issue with Pure drivers delete_snapshot exception handling

We were checking for only a single possible error that can occur when
the snapshot was missing. We now check for both which helps prevent
any snapshots or volumes deleted out from underneath Cinder to get
things stuck in an error state.

This also adjusts the warning message to be a little more descriptive.

Change-Id: I26f393a3e5232fc508b9f34921e9f53f8b5a77ef
Closes-Bug: #1553029