Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  5025c20960401f81378cc7a9e3c89c656f86dec7
Switch /etc/logstash to /etc/subunit2sql

This commit switches the use of the /etc/logstash directory to
/etc/subunit2sql. This was a leftover from when we first got the
subunit workers up and running and it was largely a copy and paste
of the policy for the logstash workers. Changing this should make it
more clear to users and consistent with the naming of everything
else.

Change-Id: Ie425dcf53856edd873ff1d0185dcf8e4ae30e3f8